+++ src/ui/vanilla/entry-point/MainSetup.ts
@@ -1,5 +1,6 @@
 import type { Widget } from "@seelen-ui/lib/types";
 import { _invoke, webviewInfo } from "src/ui/vanilla/entry-point/_tauri.ts";
+import { listen } from "@tauri-apps/api/event";
 
 import { hookLocalStorage } from "src/ui/vanilla/entry-point/_LocalStorage";
 
@@ -13,6 +14,14 @@ window.__SLU_WIDGET = widgetList.find((widget) => widget.id === currentWidgetId)
 if (!window.__SLU_WIDGET) {
   throw new Error(`Widget definition not found for ${currentWidgetId}`);
 }
+
+// reload if widget definition changed
+listen<Widget[]>("widgets-changed", ({ payload }) => {
+  const actual = payload.find((widget) => widget.id === currentWidgetId);
+  if (actual && JSON.stringify(actual) !== JSON.stringify(window.__SLU_WIDGET)) {
+    window.location.reload();
+  }
+});
 
 // set document id
 document.documentElement.id = currentWidgetId;